[FIX] DB네임, 마이그레이션명, 모델명 통일
